Lifted state from LandingPageModel.js to LandingPage.js
Because LandingPageModel is a child of LandingPage, and the NavBar is
also a child of the LandingPage, we cannot have the click handler of the
NavBar tell what the LandingPageModel to do. Therefore, we lift the
state of the 3D model LandingPageModel entirely to the LandingPage,
which is now a huge file, but this is the only way to do this nicely
given our current infrastructure.