Introduced url and state syncing.
There are still some bugs with this, but the idea is moving towards what
we want, which is to have the url and the state of the webpage sync up
so that if you refresh or open a link, the page will update to that
state. Right now, the landing page will sync with the url, but the url
is not always syncing up with the landing page. I think there can be a
simple hack where since the plus sign is how we navigate, we can just
update the url there.