Restyle line page photos so that the border is on top
Make the photos offseted from the designer name line so that there is no
need to have perfect alignment with the line.

Also added a transition delay to thge videos (1s) and shortened the
transition duration of the videos so that the play button doesn't show
in more cases, and the wait time isn't too bad.